Improved the search query
The accumulate error function was integrated int he performQuery function.
More exactly, the error is now propagated in the effect type F and not in the Either type.

Also, it uses the Fs2 Stream API to perform the search query. This means that
the return type of ScrapperResult was pretty much replaced with an F[Item]
in every file. I think this recuded the time by not having to concatenate the
errors after the query was performed, but this is only a suspicion and further
performace benchmarks are needed.